Unify index.json schema into one canonical registry-types module
The index.json shape was declared three times — indexer producer
(index_json.rs), daemon consumer (index_schema.rs), and the shared
/registry/backends leaf types — synced by comment only and drifting
(e.g. license required in the producer vs #[serde(default)] in the
consumer). Move the 9 structs + SCHEMA_VERSION/MIN_CLIENT to
super-stt-registry-types::index, deriving Serialize + Deserialize once,
and have all three sites reference it. license is now lenient on read
(default) and always written, so a slightly-off index no longer fails the
whole catalog. RegistryModel/RegistrySecret/RegistryOption/IndexStale
become aliases of the canonical leaves. The daemon's min_client soft-floor
and unsafe-path filter stay daemon-side as free functions over Index
(the validate_runtime extension pattern). Net ~450 duplicated lines gone.
